chore(canvas): tighten comments, split ActivityPanel/SpaceSidebar, drop plan doc
Cleanup pass over the spaces PR:

- Remove docs/plans/spaces-layout-prototype.md.
- Purge "Arc/Zen" wording and cut low-value comments (CSS/JSX narration, restatements, cross-file duplication); keep only non-obvious "why" and cross-file contracts.
- Split the two oversized components: ActivityPanel (696 → 457) extracts the timeline rows + node-building into ActivityTimeline.tsx; SpaceSidebar (610 → ~330) extracts the item hover-row into SpaceItemRow.tsx and the item-building into a useSpaceItems hook, and collapses the chained recent-list filters into one pass.

No behavior change. Typecheck + biome clean, 2089 UI tests pass.

Generated-By: PostHog Code
Task-Id: c401c7cb-e436-4da3-80bf-1d3008ee5d38
