+++ packages/ui/src/features/canvas/stores/spaceStore.ts
@@ -1,20 +1,17 @@
 import { create } from "zustand";
 
 /**
- * View state for the Arc-style "spaces" chrome (layout prototype): which
- * channel the sidebar is currently scoped to, which direction the last switch
- * moved (drives the slide), and the two sidebar-body overrides — browsing the
- * full channel list, or holding a draft "new space" chooser. The current space
- * survives navigating to channel-less routes (inbox, activity, settings);
- * selecting any space clears the overrides.
+ * View state for the spaces chrome: which channel the sidebar is scoped to, the
+ * last switch direction (drives the slide), and two mutually-exclusive
+ * sidebar-body overrides — `browsing` (the all-channels list) and `draftSpace`
+ * (the new-space chooser). The current space survives channel-less routes
+ * (inbox, activity, settings); selecting any space clears both overrides.
  */
 interface SpaceState {
   currentChannelId: string | null;
-  /** +1 = moved right in the dot order, -1 = left. Drives the slide-in. */
+  /** +1 = moved right in the dot order, -1 = left. */
   direction: 1 | -1;
-  /** Sidebar body shows the all-channels list (the "#" toggle). */
   browsing: boolean;
-  /** Sidebar body shows the new-space chooser (the "+" draft, Arc-style). */
   draftSpace: boolean;
   setCurrentChannel: (channelId: string | null, direction?: 1 | -1) => void;
   setBrowsing: (browsing: boolean) => void;
@@ -26,8 +23,6 @@ export const useSpaceStore = create<SpaceState>()((set) => ({
   direction: 1,
   browsing: false,
   draftSpace: false,
-  // Landing in a space always dismisses the overrides — they are transient
-  // pickers, not places.
   setCurrentChannel: (currentChannelId, direction) =>
     set((state) => ({
       currentChannelId,
